\name{negentropy}
\alias{negentropy}
\title{
Approximation of the negentropy measure.
}
\description{
It let calculate an approximation of the negentropy measure for a certain signal.
}
\usage{
negentropy(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A vector that is a signal or a stochastic variable.
}
}

\value{
It returns an float which is an approximattion of the negentropy
}


